Dev to main (#1)
* WIP - Client handler

Implement parsing the get request from client.
TODO: Craft 404 message and other error messages,
set up directory and finding requested files,
getting date time and setting expire fields in HTTP responses

* Create a file hander

needed to fille in request for pages specified by client. Takes in a path and a
request path from client, combines them. A function is aviable now to check that
the file exists. Implement a simle getter function for getting a file pointer.

* Add in function for closing server socket

Errors when compiling. Last run to test was done on void linux. The closing of a
socket file descriptor for linux and NetBSD are different. This server is more
meant to run on NetBSD, so added in compatibility and put the closing logic into
its own method

* Added a comment for documenting which argument is for what

* Commented out code that is not needed as of now

* Changed a function declaration, commented out a line of code, add include,
change function signature

Added in the file hander files
Added the addition of the root file path into the method handler
added in a test bit of code to test for the existence of requested file

* Removed and added a function, included stdio

* Added and removed a function

also included stdio

* Added file_handler header and c file to make file

* Inserted options and added more features

Addition of 'Wall' option when compiling to get errors.
Compile option for make just compile object files and places them in build file.
New link option for make will link object files into an executable

* Make use of pclose instead of close and fix preprocessor for linux

* WIP packet builder for making response packes to clients

* unneeded file

* Added new files to compile and remove during clean

* include the new WIP packet_builder

* Clean up warnings

Commented out a currently unused variable
Set file_path to null as per warning to then be set dynamically later if used
Returned null in some placed temporarily for cases not yet covered. WIP

* Fix collisions in branch

* Fix segmentation fault

seg fault was occuring due to initializing file_path as null. Fix was to keep it
initialized to null, later on calling malloc to allocation the memory, then free
the memory when it is no longer needed

* Update README.md

* Added in functionality to response with 404 from server when client asks for a resource that does not exist.

* Added the signature for method to get file length

* Implement function to get the file length

* Fixed the offset

* AMake and implement a method to get file contents

i Please enter the commit message for your changes. Lines starting

* Implement the method to get file contents

* Add a comment regarding the purpose of the function

* Added string.h to include, removed a function decided that it was not needed and
added function signature for a function that will make the http packet into a
string to send to client

* Implement making a string representation of a packet in a function

* Use new functionality to craft a packet and get a string represenation of a packet