Dev to main: Implement sending images and CSS (#9)
* WIP - Client handler

Implement parsing the get request from client.
TODO: Craft 404 message and other error messages,
set up directory and finding requested files,
getting date time and setting expire fields in HTTP responses

* Create a file hander

needed to fille in request for pages specified by client. Takes in a path and a
request path from client, combines them. A function is aviable now to check that
the file exists. Implement a simle getter function for getting a file pointer.

* Add in function for closing server socket

Errors when compiling. Last run to test was done on void linux. The closing of a
socket file descriptor for linux and NetBSD are different. This server is more
meant to run on NetBSD, so added in compatibility and put the closing logic into
its own method

* Added a comment for documenting which argument is for what

* Commented out code that is not needed as of now

* Changed a function declaration, commented out a line of code, add include,
change function signature

Added in the file hander files
Added the addition of the root file path into the method handler
added in a test bit of code to test for the existence of requested file

* Removed and added a function, included stdio

* Added and removed a function

also included stdio

* Added file_handler header and c file to make file

* Inserted options and added more features

Addition of 'Wall' option when compiling to get errors.
Compile option for make just compile object files and places them in build file.
New link option for make will link object files into an executable

* Make use of pclose instead of close and fix preprocessor for linux

* WIP packet builder for making response packes to clients

* unneeded file

* Added new files to compile and remove during clean

* include the new WIP packet_builder

* Clean up warnings

Commented out a currently unused variable
Set file_path to null as per warning to then be set dynamically later if used
Returned null in some placed temporarily for cases not yet covered. WIP

* Fix collisions in branch

* Fix segmentation fault

seg fault was occuring due to initializing file_path as null. Fix was to keep it
initialized to null, later on calling malloc to allocation the memory, then free
the memory when it is no longer needed

* Update README.md

* Added in functionality to response with 404 from server when client asks for a resource that does not exist.

* Added the signature for method to get file length

* Implement function to get the file length

* Fixed the offset

* AMake and implement a method to get file contents

i Please enter the commit message for your changes. Lines starting

* Implement the method to get file contents

* Add a comment regarding the purpose of the function

* Added string.h to include, removed a function decided that it was not needed and
added function signature for a function that will make the http packet into a
string to send to client

* Implement making a string representation of a packet in a function

* Use new functionality to craft a packet and get a string represenation of a packet

* WIP Implementing packet parser

Work in progress. Working on the parser for HTTP requests from client.

* Implement the first complete iteration of the packet parser

Only handles GET requests currently

* Remove unneeded test code

* Added some roadmap goals to README

* Change formating and added another roadmap goal

* Remove unneeded variables

since only one delimiter is used, rename it to a generic delim. Remove the space
delim and the unused token variable

* Update readme to reflect the layout of respository

* Implement sending handling files that exist

Sends proper http packets back to client for resources that exist. However, WIP
is handling multiple different requests.

* Add a TODO

this todo right now capture items needed for release 1.0

* Remove some code for future testing

* Fix implementation for serving multiple back to back requests

Still needs to be tested. The message buffer going out was not being properly
cleared. Also freeing memory sent out to a function for a more clean look.

* Added flag

Flag added for potential debugging options

* Cross of item from todo list

* Removed commented out code that is no longer needed

* Add comments for documentation of functions

* Removed unneeded test code

* Add comments

* Removed lines for test

* Remove unneeded code

* Use pclose to satify compile warning

* Implement default request

if the requested resource is "/" then it will default it to "/index.html"

* Change the name of the output program

* Fix formatting

* Change formatting

Added in checkboxes to list to track along with what release point is completed

* Remove link rule

Removed the link rule and just placed the call for the linker to link the files
under the compile rule

* Rule for CI

Jenkins failed build after combining the original compile rule with the link
rule. Now creating a special rule for jenkins to run.

* Implement TRUE

Defined TRUE to be 1. This is to improve readability in the while loop.

* Add directive to include unistd

Clear compiler warning for implicit declaration of 'write.'

* Update TODO to match 2.0 release goals

* Mark release one complete and change roadmap for release 4

Release 4 will now target handling a config file

* Function definition added

Function definition added for a function that takes a path and returns the extension

* Implement get_file_extentsion

* Added completed TODO item

* Add a function declaration

New function declaration declares the signature of a function that will look at
an extension and return the corresponding content type for an HTTP packet

* Implement get_content_type function

Implement a function that takes in an extension and returns the corresponding
HTTP packet content type for that data to be sent

* WIP - handling sending files

Issue with sending files over HTTP. Needed to send images as binary. Not issue
with message body. For potential solution, look into unions to define the type
of message body?

* Change to roadmap

* Made message body void pointer to accept any kind of data for message body

* Update TODO.md

Added a TODO item for using valgrind to find memory leaks and then patch

* WIP putting binary data in packet

* Implemented sending images over http and other fixes

Make a correction to the shutdown for linux. Previous solution worked on Void,
however not on Debian.

commented out embedding the requested data in the packet, now the program writes
twice to clients. First write is for header. Next write if for requested file data.

* changes to satisfy magit

* Add css type to get_content_type

* Updated

Server can not return images alongside CSS files

Co-authored-by: Todd Martin <todd@debian>